% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{typdb.data.merge}
\alias{typdb.data.merge}
\title{Retrieve and merge TypDB datasets}
\usage{
typdb.data.merge(..., result = c("data", "mapping", "index"))
}
\arguments{
\item{...}{Two or more TypDB dataset names}

\item{result}{An option controlling what to return (see details)}
}
\value{
A data frame that is either the merged datasets, the language code
mapping table or the join table indicating which rows will be joined
(depending on \code{result} parameter)
}
\description{
\code{typdb.data.merge} merges data from different TypDB
datasets based on the overlapping language codes (see \link{language_ids}).
}
\details{
The \code{result} parameter controls what the function will return.
If the value of this parameter is \code{"data"} (default), the merged data
frame is returned. If \code{result} is \code{"mapping"}, the language
mappings required to perform the merge are returned instead (this is a
relevant subset of \link{language_ids} data frame). Finally, if \code{result}
is \code{"index"}, the result is a merged data frame consisting of row
numbers of the datasets that are merged together. This is useful in trackign
down any language code mismatches.

The function will produce a warning if any language codes are duplicated,
missing, or if there is an ambigues match. Furthermore, the column names of
the resulting data frame will be disambiguated if nessesary.
}

