% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate_dataset.R
\name{typdb.locate.dataset}
\alias{typdb.locate.dataset}
\title{Finds TypDB datasets}
\usage{
typdb.locate.dataset(language.code, result = c("data", "names"))
}
\arguments{
\item{language.code}{A characer vector of language codes (of any type)}

\item{result}{Controls what is returned}
}
\value{
If \code{result} is \code{"data"}, a list of with appropriate data
frames. If \code{result} is \code{"names"}, a character vectors of dataset
names.
}
\description{
Finds all the datasets in TypDB for with data on given languages
}

